requirements.txt
numpy>=1.17.0
requests>=2.25.1
torchvision>=0.6.0
yacs>=0.1.7
kornia>=0.3.1
matplotlib>=3.3.4
torch>=1.5.0
glog>=0.3.1
gradio>=1.6.4
seaborn>=0.11.0
Pillow>=8.2.0
scikit_learn>=0.24.1
wandb
lpips